+++ packages/javascript/lumis/tsup.config.ts
@@ -31,7 +31,6 @@ export default defineConfig({
     index: "src/index.ts",
     "index.browser": "src/index.browser.ts",
     cache: "src/cache.ts",
-    "cache-cli": "src/cache-cli.ts",
     formatters: "src/formatters.ts",
     "formatters/html": "src/formatter/html.ts",
     "formatters/ansi": "src/formatter/ansi.ts",